requirements.txt
# Dependências para o Extrator de Orçamentos PDF para Google Sheets

# Interface gráfica para seleção de arquivos
# tkinter já vem incluído com Python

# Extração de texto de PDFs
PyMuPDF==1.23.8

# Integração com Google Sheets
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Processamento de dados e regex
# re já vem incluído com Python

# Manipulação de datas
# datetime já vem incluído com Python

# Processamento de JSON
# json já vem incluído com Python
